Exclude google-payment from braintree test dependency
In a merchant application braintree and google-payment dependencies
would be defined:

    // Note: braintree:2.19.0+ includes google-payment:1.0.0
    implementation 'com.braintreepayments.api:braintree:2.19.0'
    //         +--- com.braintreepayments.api:google-payment:1.0.0

    implementation 'com.braintreepayments.api:google-payment:2.0.0'

Gradle would detect that two google-payment dependencies are requested:
1.0.0, and 2.0.0. It should resolve with the higher version.

For running google-payment unit tests Gradle is only detecting
google-payment:1.0.0 via `./gradlew :GooglePayment:dependencies`

I believe this is due to no higher version being available in the source
repositories. Gradle is pulling in the local working copy of
google-payment which it doesn't seem to perform dependency resolution on
it.

The solution is to exclude the google-payment module for testing.

This is the same solution we take for braintree_android, when asking
Drop-in to use the local braintree_android version in the
Demo/build.gradle:
https://github.com/braintree/braintree_android/blob/master/Demo/build.gradle#L94-L96