Add check in `requestPayment` to avoid NPE
This fix will return an exception if `requestPayment` is called when
Google Pay is not enabled.  Ideally this would be caught in
`isReadyToPay`, however it's possible though not advised to jump
straight to calling `requestPayment` and get a Null Pointer Exception if
some values are not set in the configuration.