Bump core SDK version. Fix broken tests
This bumps the core SDK dependency version to the current version to
include 3DSecure, in the event it's required.

It also fixes an issue where the wrong SDK version was being passed to
various fields.  This was coincidentally correct as the base SDK version
matched the version of the google pay module.  When the no longer
matched, the bug presented itself.